+++ next-docs/components/facing/CardArticle.tsx
@@ -1,6 +1,4 @@
 import React from 'react';
-import { rem } from '@heathmont/moon-utils';
-import Link from 'next/link';
 import classNames from '../../utils/classNames';
 
 type Props = {
@@ -12,18 +10,9 @@ type Props = {
 };
 
 const variants = {
-  default: {
-    width: rem(384),
-    height: rem(288),
-  },
-  tall: {
-    width: rem(288),
-    height: rem(448),
-  },
-  wide: {
-    width: rem(448),
-    height: rem(288),
-  },
+  tall: 'md:w-[18rem] md:h-[28rem] md:pe-9',
+  wide: 'md:w-[28rem] md:h-[18rem] md:pe-12',
+  default: 'md:w-[24rem] md:h-[18rem] md:pe-12',
 };
 
 const CardArcticle: React.FC<Props> = ({
@@ -32,17 +21,14 @@ const CardArcticle: React.FC<Props> = ({
   section,
   href,
   author,
-}) => (
-  <Link href={href}>
-    <a target="_blank">
+}) => {
+  return (
+    <a href={href} target="_blank" rel="noreferrer">
       <div
         className={classNames(
-          variant === 'tall' ? 'pe-9' : 'pe-12',
-          `group relative flex flex-col justify-end ps-6 py-6 bg-goku rounded-moon-s-lg`
-        )}
-        style={
+          `group relative w-72 h-96 flex flex-col justify-end p-6 bg-goku rounded-moon-s-lg`,
           variant && variant in variants ? variants[variant] : variants.default
-        }
+        )}
       >
         <p className="list-item absolute top-10 ltr:left-10 rtl:right-10 transform -rotate-90 origin-bottom-left text-moon-16 font-medium">
           {section}
@@ -55,7 +41,7 @@ const CardArcticle: React.FC<Props> = ({
         <p className="text-moon-24 font-medium">{title}</p>
       </div>
     </a>
-  </Link>
-);
+  );
+};
 
 export default CardArcticle;