+++ next-docs/components/facing/ForDevelopers.tsx
@@ -22,49 +22,51 @@ const GithubIcon = () => (
 
 const ForDevelopers = () => {
   return (
-    <div className="relative flex grow p-6 gap-6 theme-tokens rounded-moon-s-lg bg-cell 3xl:flex-col 3xl:h-96 3xl:w-80 ltr:3xl:rounded-tl-none rtl:3xl:rounded-tr-none 3xl:absolute 3xl:top-96 ltr:3xl:right-0 rtl:3xl:left-0">
+    <div className="relative z-50 flex grow p-6 gap-6 theme-tokens rounded-moon-s-lg bg-cell 3xl:flex-col 3xl:h-96 3xl:w-80 ltr:3xl:rounded-tl-none rtl:3xl:rounded-tr-none 3xl:absolute 3xl:top-96 ltr:3xl:right-0 rtl:3xl:left-0">
       <div className="flex gap-6 align-center justify-between grow 3xl:flex-col 3xl:justify-end">
         <GithubIcon />
-        <h2 className="text-moon-24 grow 3xl:grow-0">
-          <a
-            href="https://github.com/coingaming/moon-design"
-            target="_blank"
-            rel="noreferrer"
-          >
-            For developers.
-          </a>
-        </h2>
-        <div className="3xl:absolute 3xl:top-6 ltr:3xl:right-6 rtl:3xl:left-6 rtl:-rotate-90">
-          <TopRightArrowIcon />
+        <h2 className="text-moon-24 grow 3xl:grow-0">For developers.</h2>
+        <div className="absolute top-0 bottom-0 left-0 right-0 3xl:top-6 3xl:bottom-auto ltr:3xl:right-6 ltr:3xl:left-auto rtl:3xl:left-6 rtl:3xl:right-auto rtl:-rotate-90">
+          <Link href="/gettingStartedDevs">
+            <a className="flex items-center justify-end h-full pe-6 3xl:pe-0 w-full rounded-moon-s-lg">
+              <TopRightArrowIcon />
+            </a>
+          </Link>
         </div>
       </div>
       <div className="hidden flex-wrap gap-1 3xl:flex">
-        <Link href="https://github.com/coingaming/moon-design">
-          <a target="_blank" rel="noreferrer">
-            <Button variant="secondary">Github</Button>
-          </a>
-        </Link>
-        <Link href="gettingStarted">
+        <a
+          href="https://github.com/coingaming/moon-design"
+          target="_blank"
+          rel="noreferrer"
+        >
+          <Button variant="secondary">Github</Button>
+        </a>
+        <Link href="/gettingStartedDevs">
           <a>
             <Button variant="secondary">Getting Started</Button>
           </a>
         </Link>
-        <Link href="tokens">
+        <Link href="/tokens">
           <a>
             <Button variant="secondary">Design Tokens</Button>
           </a>
         </Link>
-        <Link href="https://github.com/coingaming/moon-design/blob/develop/packages/themes/src/moonDesignDark/moonDesignDark.ts">
-          <a target="_blank" rel="noreferrer">
-            <Button variant="secondary">Creating Themes</Button>
-          </a>
-        </Link>
-        <Link href="https://github.com/coingaming/moon-design/blob/develop/CONTRIBUTING.md">
-          <a target="_blank" rel="noreferrer">
-            <Button variant="secondary">Contribution</Button>
-          </a>
-        </Link>
-        <Link href="components/accordion">
+        <a
+          href="https://github.com/coingaming/moon-design/blob/develop/packages/themes/src/moonDesignDark/moonDesignDark.ts"
+          target="_blank"
+          rel="noreferrer"
+        >
+          <Button variant="secondary">Creating Themes</Button>
+        </a>
+        <a
+          href="https://github.com/coingaming/moon-design/blob/develop/CONTRIBUTING.md"
+          target="_blank"
+          rel="noreferrer"
+        >
+          <Button variant="secondary">Contribution</Button>
+        </a>
+        <Link href="/core/accordion">
           <a>
             <Button variant="secondary">Components</Button>
           </a>