+++ next-docs/pages/components/charts/geoMapChart.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import Link from 'next/link';
 import Preview from '../../../components/codePreview/Preview';
 import CodeSnippet from '../../../components/CodeSnippet';
 import ComponentPageDescription from '../../../components/ComponentPageDescription';
@@ -15,25 +14,23 @@ export default function PageGeoMapChart() {
       <ComponentPageDescription title="Geo map">
         <p>
           Based on{' '}
-          <Link href="https://www.react-simple-maps.io/">
-            <a
-              className="text-piccolo font-medium transition-colors duration-200 hover:text-hit visited:text-hit"
-              target="_blank"
-              rel="noreferrer"
-            >
-              react-simple-maps
-            </a>
-          </Link>{' '}
+          <a
+            href="https://www.react-simple-maps.io/"
+            className="text-piccolo font-medium transition-colors duration-200 hover:text-hit visited:text-hit"
+            target="_blank"
+            rel="noreferrer"
+          >
+            react-simple-maps
+          </a>{' '}
           and{' '}
-          <Link href="https://recharts.org/en-US/">
-            <a
-              className="text-piccolo font-medium transition-colors duration-200 hover:text-hit visited:text-hit"
-              target="_blank"
-              rel="noreferrer"
-            >
-              Recharts
-            </a>
-          </Link>
+          <a
+            href="https://recharts.org/en-US/"
+            className="text-piccolo font-medium transition-colors duration-200 hover:text-hit visited:text-hit"
+            target="_blank"
+            rel="noreferrer"
+          >
+            Recharts
+          </a>
           .
         </p>
       </ComponentPageDescription>