+++ next-docs/pages/components/charts/verticalBarChart.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import Link from 'next/link';
 import Preview from '../../../components/codePreview/Preview';
 import CodeSnippet from '../../../components/CodeSnippet';
 import ComponentPageDescription from '../../../components/ComponentPageDescription';
@@ -16,15 +15,14 @@ export default function PageVerticalBarChart() {
       <ComponentPageDescription title="Vertical bar">
         <p>
           Based on{' '}
-          <Link href="https://recharts.org/en-US/">
-            <a
-              className="text-piccolo font-medium transition-colors duration-200 hover:text-hit visited:text-hit"
-              target="_blank"
-              rel="noreferrer"
-            >
-              Recharts
-            </a>
-          </Link>
+          <a
+            href="https://recharts.org/en-US/"
+            className="text-piccolo font-medium transition-colors duration-200 hover:text-hit visited:text-hit"
+            target="_blank"
+            rel="noreferrer"
+          >
+            Recharts
+          </a>
           .
         </p>
       </ComponentPageDescription>