+++ next-docs/pages/components/table.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import Link from 'next/link';
 import Preview from '../../components/codePreview/Preview';
 import ComponentPageDescription from '../../components/ComponentPageDescription';
 import PropsTable from '../../components/PropsTable';
@@ -24,15 +23,14 @@ export default function PageTable() {
         <p>Lightweight, fast and extendable table.</p>
         <p>
           Based on{' '}
-          <Link href="https://github.com/tannerlinsley/react-table">
-            <a
-              className="text-piccolo font-medium transition-colors duration-200 hover:text-hit visited:text-hit"
-              target="_blank"
-              rel="noreferrer"
-            >
-              react-table
-            </a>
-          </Link>
+          <a
+            href="https://github.com/tannerlinsley/react-table"
+            className="text-piccolo font-medium transition-colors duration-200 hover:text-hit visited:text-hit"
+            target="_blank"
+            rel="noreferrer"
+          >
+            react-table
+          </a>
           .
         </p>
       </ComponentPageDescription>