+++ next-docs/pages/core/radiobutton.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import Preview from '../../components/codePreview/Preview';
 import ComponentPageDescription from '../../components/ComponentPageDescription';
+import PropsTable from '../../components/PropsTable';
 import Checked from '../../public/examples/radiobuttonTW/Checked';
 import Default from '../../public/examples/radiobuttonTW/Default';
 import Disabled from '../../public/examples/radiobuttonTW/Disabled';
@@ -49,6 +50,67 @@ const Example = () => {
         title="ReadOnly"
         preview={<ReadOnly />}
         code={examples ? examples.ReadOnly : 'Loading'}
+      />
+      <PropsTable
+        title="RadioButton props"
+        data={[
+          {
+            name: 'ariaLabel',
+            type: 'string',
+            required: false,
+            default: '-',
+            description: 'ARIA description of radio button',
+          },
+          {
+            name: 'checked',
+            type: 'boolean',
+            required: false,
+            default: 'false',
+            description: 'Checked state of radio button',
+          },
+          {
+            name: 'disabled',
+            type: 'boolean',
+            required: false,
+            default: 'false',
+            description: 'Disables radio button',
+          },
+          {
+            name: 'id',
+            type: 'string',
+            required: false,
+            default: '-',
+            description: `Sets radio button id`,
+          },
+          {
+            name: 'label',
+            type: 'JSX.Element | string',
+            required: false,
+            default: '-',
+            description: 'Label for radio button',
+          },
+          {
+            name: 'onChange',
+            type: '() => void',
+            required: false,
+            default: '-',
+            description: 'Change event handler',
+          },
+          {
+            name: 'onClick',
+            type: '() => void',
+            required: false,
+            default: '-',
+            description: 'Click event handler',
+          },
+          {
+            name: 'readOnly',
+            type: 'boolean',
+            required: false,
+            default: 'false',
+            description: "Value of radio button can't be changed",
+          },
+        ]}
       />
     </>
   );