+++ next-docs/public/examples/labelTW/Colours.tsx
@@ -3,10 +3,10 @@ import { Label } from '@heathmont/moon-core-tw';
 const Example = () => (
   <>
     <Label>Xsmall</Label>
-    <Label color="text-goten" backgroundColor="bg-trunks">
+    <Label color="text-goten" bgColor="bg-roshi">
       Xsmall
     </Label>
-    <Label color="text-goten" backgroundColor="bg-dodoria">
+    <Label color="text-goten" bgColor="bg-dodoria">
       Xsmall
     </Label>
   </>