+++ workspaces/core/src/radioButton/private/types/Props.ts
@@ -4,6 +4,7 @@ type Props = React.InputHTMLAttributes<HTMLInputElement> & {
   label?: JSX.Element | string;
   id?: string;
   checked?: boolean;
+  readOnly?: boolean;
   onClick?: (event?: React.MouseEvent<HTMLInputElement>) => void;
   onChange?: (event: React.ChangeEvent<HTMLInputElement>) => void;
 };