+++ workspaces/core/src/textInput/private/types/TextInputProps.ts
@@ -4,14 +4,14 @@ export default interface TextInputProps
   extends React.InputHTMLAttributes<HTMLInputElement> {
   id?: string;
   inputSize?: 'md' | 'lg' | 'xl';
-  type: TextInputTypes | string;
+  type?: TextInputTypes | string;
   label?: JSX.Element | string;
   placeholder?: string;
   hintText?: JSX.Element | string;
   isError?: boolean;
   dir?: 'ltr' | 'rtl' | 'auto';
   showPasswordText?: JSX.Element | string;
-  backgroundColor?: string;
+  bgColor?: string;
   isSharpLeftSide?: boolean;
   isSharpRightSide?: boolean;
   isSharpTopSide?: boolean;