+++ workspaces/tables/src/styled/StyledTable.ts
@@ -1,0 +1,9 @@
+export const table = 'flex flex-col gap-1 rounded-moon-s-sm p-1 bg-goku';
+export const trHeader =
+  'flex gap-4 rounded-moon-s-sm bg-goku text-bulma moon-text-14 px-4 mb-1';
+export const trBody =
+  'flex gap-4 rounded-moon-s-sm bg-gohan hover:bg-piccolo/[.12] transition-colors text-bulma moon-text-14 px-4 mb-1';
+export const td = 'py-4';
+export const th = 'py-3';
+
+export default { table, trHeader, trBody, td, th };